+++ client.go
@@ -36,6 +36,7 @@ type Client struct {
 	LoadBalancers   *LoadBalancerService
 	ObjectStorages  *ObjectStorageService
 	Servers         *ServerService
+	ServerDisks     *ServerDisksService
 	Networks        *NetworkService
 	NetworkAdapters *NetworkAdapterService
 }
@@ -59,6 +60,7 @@ func NewClient(project, apiKey, userAgent string) *Client {
 	c.LoadBalancers = &LoadBalancerService{client: c}
 	c.ObjectStorages = &ObjectStorageService{client: c}
 	c.Servers = &ServerService{client: c}
+	c.ServerDisks = &ServerDisksService{client: c}
 	c.Networks = &NetworkService{client: c}
 	c.NetworkAdapters = &NetworkAdapterService{client: c}